+++ R/extractors.R
@@ -1,8 +1,8 @@
-gold_classes <- function(x) {
-    UseMethod("gold_classes")
+get_gold_classes <- function(x) {
+    UseMethod("get_gold_classes")
 }
 
-gold_classes.goldrake <- function(x) {
+get_gold_classes.goldrake <- function(x) {
     attr(x, "gold_classes")
 }
 
@@ -12,11 +12,11 @@ gold_classes.goldrake <- function(x) {
 
 
 
-reviewer_names <- function(x) {
-    UseMethod("reviewer_names")
+get_reviewer_names <- function(x) {
+    UseMethod("get_reviewer_names")
 }
 
-reviewer_names.goldrake <- function(x) {
+get_reviewer_names.goldrake <- function(x) {
     attr(x, "reviewer_names")
 }
 
@@ -24,14 +24,38 @@ reviewer_names.goldrake <- function(x) {
 
 
 
-
-
-
-balanced_variables <- function(x) {
-    UseMethod("balanced_variables")
+get_original_data <- function(x) {
+    UseMethod("get_original_data")
 }
 
-balanced_variables.goldrake <- function(x) {
+get_original_data.goldrake <- function(x) {
+    x[["original_data"]]
+}
+
+
+
+
+
+get_used_data <- function(x) {
+    UseMethod("get_used_data")
+}
+
+get_used_data.goldrake <- function(x) {
+    x[["used_data"]]
+}
+
+
+
+
+
+
+
+
+get_balanced_variables <- function(x) {
+    UseMethod("get_balanced_variables")
+}
+
+get_balanced_variables.goldrake <- function(x) {
     attr(x, "balanced_variables")
 }
 
@@ -42,11 +66,11 @@ balanced_variables.goldrake <- function(x) {
 
 
 
-max_sampled <- function(x) {
-    UseMethod("max_sampled")
+get_max_sampled <- function(x) {
+    UseMethod("get_max_sampled")
 }
 
-max_sampled.goldrake <- function(x) {
+get_max_sampled.goldrake <- function(x) {
     attr(x, "max_sampled")
 }